Add custom allocator and dynamic array
Custom allocator adds capability to define custom allocator with the
given interface. This interface is used throughout the library,
specially dynamic array.

The dynamic array is aware of the allocator state, which allows to swap
out the allocator to choose different one. This allows to create temp
allocator that does fast allocation and avoids heap all together.

The allocator scope swaps the global allocator with the provided custom
allocator.